Extract duplicated getPublicReadableRealms into shared utility
Move the identical getPublicReadableRealms function from both
handle-federated-types.ts and handle-realm-info.ts into
utils/realm-readability.ts to eliminate duplication.
