+++ packages/realm-server/handlers/handle-federated-types.ts
@@ -1,18 +1,10 @@
 import type Koa from 'koa';
 import type { DBAdapter } from '@cardstack/runtime-common';
-import {
-  ensureTrailingSlash,
-  fetchUserPermissions,
-  logger,
-  SupportedMimeType,
-} from '@cardstack/runtime-common';
+import { logger, SupportedMimeType } from '@cardstack/runtime-common';
 import { makeCardTypeSummaryDoc } from '@cardstack/runtime-common/document-types';
 import { setContextResponse } from '../middleware';
 import { getMultiRealmAuthorization } from '../middleware/multi-realm-authorization';
-import {
-  buildReadableRealms,
-  getPublishedRealmURLs,
-} from '../utils/realm-readability';
+import { getPublicReadableRealms } from '../utils/realm-readability';
 
 const log = logger('realm-server');
 
@@ -30,23 +22,17 @@ export default function handleFederatedTypes({
 
     let data: Record<string, ReturnType<typeof makeCardTypeSummaryDoc>> = {};
 
-    let results = await Promise.allSettled(
-      realmList.map(async (realmURL) => {
-        let realm = realmByURL.get(realmURL);
-        if (!realm) {
-          return;
-        }
+    for (let realmURL of realmList) {
+      let realm = realmByURL.get(realmURL);
+      if (!realm) {
+        continue;
+      }
+      try {
         let summaries =
           await realm.realmIndexQueryEngine.fetchCardTypeSummary();
         data[realmURL] = makeCardTypeSummaryDoc(summaries);
-      }),
-    );
-
-    for (let [index, result] of results.entries()) {
-      if (result.status === 'rejected') {
-        log.warn(
-          `Failed to fetch card type summary for realm ${realmList[index]}: ${result.reason}`,
-        );
+      } catch (error) {
+        log.warn(`Failed to fetch card type summary for ${realmURL}: ${error}`);
       }
     }
 
@@ -63,27 +49,4 @@ export default function handleFederatedTypes({
       new Response(JSON.stringify({ data }, null, 2), { headers }),
     );
   };
-}
-
-async function getPublicReadableRealms(
-  dbAdapter: DBAdapter,
-  realmList: string[],
-): Promise<Set<string>> {
-  let publicPermissions = await fetchUserPermissions(dbAdapter, {
-    userId: '*',
-    onlyOwnRealms: false,
-  });
-
-  let publishedRealmURLs = await getPublishedRealmURLs(dbAdapter, realmList);
-  let publicReadable = buildReadableRealms(
-    publicPermissions,
-    publishedRealmURLs,
-  );
-
-  let normalizedRealmList = realmList.map((realmURL) =>
-    ensureTrailingSlash(realmURL),
-  );
-  return new Set(
-    normalizedRealmList.filter((realmURL) => publicReadable.has(realmURL)),
-  );
 }
